+++ example/lib/config/api.dart
@@ -1,4 +1,7 @@
 Map<String, dynamic> apiConfig = {
   /// Api service base url
-  'url': 'https://example.com/api/'
+  'url': 'https://example.com/api/',
+
+  /// Api service create path
+  'create': ''
 };