+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [0.0.3] - Fixing dependencies versions.
+
+- Updated dependencies.
+- Removed `Validator` package.
+- Added `Validator` package content in the magic library.
+- Fixed booting problem about auth.
+- Added value control for `DatePicker` widget.
+- Added `ImagePicker` widget.
+- Added `image_picker` package.
+- Added `image_cropper` package.
+
 ## [0.0.2] - Fixing first release issues
 
 - Added license.